+++ PPO/README.txt
@@ -1,1 +1,7 @@
-pip install phevaluator
+# Libraries Used (necessary)
+phevaluator - hand strength evaluation
+torch - neuronetwork
+numpy - data calculation
+
+# Libraries Used (can be omitted)
+tensorboard - testing only
